Copy newVNode in renderComponent (+16 B)
This change fixes the problem described in the removed comments. `getDomSibling` relies on the `_parent` and `_children` relationship. But copying oldVNode in renderComponent broke that relationship. So instead this commit copies oldVNode into newVNode, so the relationship between oldVNode and it's parent is retained. Once the diff is complete, we replace oldVNode with newVNode in it's parent's children array so newVNode persists in new renders